+++ branches/1.2/src/crprober/src/main.cpp
@@ -293,7 +293,7 @@ int process_report(LPTSTR szInput, LPTSTR szInputMD5, LPTSTR szOutput,
   
   // Decide input dir and file name
   sInDirName = szInput;
-  int pos = sInDirName.rfind('\\');
+  size_t pos = sInDirName.rfind('\\');
   if(pos<0) // There is no back slash in path
   {
     sInDirName = _T(""); 
@@ -331,7 +331,7 @@ int process_report(LPTSTR szInput, LPTSTR szInputMD5, LPTSTR szOutput,
     if(bInputMD5FromDir)
     {
       sMD5DirName = szInputMD5;
-      int pos = sMD5DirName.rfind('\\');
+      size_t pos = sMD5DirName.rfind('\\');
       if(pos<0) // There is no back slash in path
         sMD5DirName = _T(""); 
       else if(pos!=(int)sMD5DirName.length()-1) // Append the back slash to dir name
@@ -758,7 +758,7 @@ int output_document(CrpHandle hReport, FILE* f)
 
         if(!sSourceFile.empty())
         {
-          int pos = sSourceFile.rfind('\\');
+          size_t pos = sSourceFile.rfind('\\');
           if(pos>=0)
             sSourceFile = sSourceFile.substr(pos+1);
           str += _T(" [ ");
